Add discharge visualization
This commit adds one new command and two new variables.

- dischargeradius
- r_dischargeLight
- r_dischargeLightColor

The dischargeradius command lets you visualize the radius of a
discharge. The color of the sphere can be controlled by setting
r_dischargeLightColor.

The discharge sphere can also be rendered automatically when playing on
a KTX server that emits the //ktx discharge event. To enable this, set
r_dischargeLight to 1.